Fixing issue with leaving dirty input forms
There are forms on UI that can be left in dirty state by using main
menu - West menu.
This was solved by adding dirty form flag on gwtSession object and
when user uses West menu it is notified that changes will be lost and
can decide if this is ok or not.

Changes were required to WestNavigationView and that could break unexpected
functionality of UI. It was tested but not overall UI tested.

This fixes issue #1137.

Signed-off-by: Uros Mesaric Kunst <uros.mesaric-kunst@comtrade.com>